requirements.txt
requests>=2.28
pandas>=1.4
matplotlib>=3.5
python-dotenv>=1.0
Flask>=2.3
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6.2
APScheduler>=3.10.0
Werkzeug>=2.3
